refactor: abstract out components
